+++ packages/declaration/src/assets/js/utils.js
@@ -234,6 +234,66 @@ checkDate = event => {
   target.reportValidity()
 }
 
+checkDatePost2018 = event => {
+  const target = event.target
+
+  checkPatternValidity(event)
+  if (target.validity.patternMismatch) {
+    // We already treated this case in `checkPatternValidity`
+    return
+  }
+
+  if (target.validity.valueMissing) {
+    // Keep the default browser behavior
+    return
+  }
+
+  const parsedDate = new Date(target.value).toString()
+  if (parsedDate === "Invalid Date") {
+      // We check if the length is >= 2 because the list of sirens also contains the current value
+      target.setCustomValidity("Veuillez saisir une date valide au format aaaa-mm-jj (exemple : 2021-11-23)")
+    } else if (new Date(target.value).getFullYear() < 2018) {
+    target.setCustomValidity("La date de consultation du CSE doit être postérieure ou égale à 2018")
+  } else {
+    target.setCustomValidity("")
+  }
+  target.reportValidity()
+}
+
+checkDatePublication = (event) => {
+    const target = event.target;
+    const min = event.target.min;
+
+    checkPatternValidity(event);
+    if (target.validity.patternMismatch) {
+        // We already treated this case in `checkPatternValidity`
+        return;
+    }
+
+    if (target.validity.valueMissing) {
+        // Keep the default browser behavior
+        return;
+    }
+
+    const parsedDate = new Date(target.value);
+    const minDate = new Date(min);
+
+    if (parsedDate.toString() === "Invalid Date") {
+        // We check if the length is >= 2 because the list of sirens also contains the current value
+        target.setCustomValidity(
+            "Veuillez saisir une date valide au format aaaa-mm-jj (exemple : 2021-11-23)"
+        );
+    } else if (parseDate < minDate) {
+        target.setCustomValidity(
+            `La date ne peut précéder la date de fin de la période de référence choisie pour le calcul de votre index (${min.toLocaleDateString(
+                "fr-FR"
+            )})`
+        );
+    } else {
+        target.setCustomValidity("");
+    }
+};
+
 extractKey = flatKey => {
   // This extracts "foobar[0]" into ["foobar[0]", "foobar", "0"]
   return flatKey.match(/([^\[]+)\[?(\d+)?\]?/);